Push failing tests for @AssistFactory field inject
