--- blackbox-test-inject/src/main/java/org/example/myapp/assist/OtherThing.java
@@ -1,19 +1,0 @@
-package org.example.myapp.assist;
-
-import io.avaje.inject.Component;
-import org.example.myapp.assist.css.OtherScanner$AssistFactory;
-
-@Component
-public class OtherThing {
-
-  private final OtherScanner$AssistFactory factory;
-
-  OtherThing(OtherScanner$AssistFactory factory) {
-    this.factory = factory;
-  }
-
-  public String scan(String file) {
-    Scanner scanner = factory.create(file);
-    return scanner.scan();
-  }
-}
