+++ blackbox-test-inject/src/main/java/org/example/myapp/assist/driod/DroidFactory.java
@@ -1,0 +1,10 @@
+package org.example.myapp.assist.driod;
+
+public interface DroidFactory {
+
+  Droid createDroid(int personality, Model model);
+
+  interface Droid {
+
+  }
+}
