+++ blackbox-test-inject/src/main/java/org/example/myapp/assist/driod/DroidMethod.java
@@ -1,0 +1,22 @@
+package org.example.myapp.assist.driod;
+
+import io.avaje.inject.Assisted;
+import io.avaje.inject.AssistFactory;
+import jakarta.inject.Inject;
+import jakarta.inject.Named;
+
+@Named("androidmethod")
+@AssistFactory(DroidFactory.class)
+public class DroidMethod implements DroidFactory.Droid {
+
+  private int personality;
+  private Model model;
+  private Engine wiring;
+
+  @Inject
+  public void inject(@Assisted int personalityMatrix, @Assisted Model model, Engine wiring) {
+    this.personality = personalityMatrix;
+    this.model = model;
+    this.wiring = wiring;
+  }
+}
