+++ blackbox-test-inject/src/main/java/org/example/myapp/assist/driod/Paint.java
@@ -1,0 +1,3 @@
+package org.example.myapp.assist.driod;
+
+public interface Paint {}
