--- inject-generator/src/test/java/io/avaje/inject/generator/models/valid/assist/Paint.java
@@ -1,3 +1,0 @@
-package io.avaje.inject.generator.models.valid.assist;
-
-public interface Paint {}
