+++ inject/src/main/java/io/avaje/inject/Assisted.java
@@ -7,7 +7,10 @@ import static java.lang.annotation.RetentionPolicy.SOURCE;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-/** */
+/**
+ * Used with {@link AssistFactory} to mark the properties that are
+ * provided as parameters to the factory method.
+ */
 @Target({FIELD, PARAMETER})
 @Retention(SOURCE)
 public @interface Assisted {}
